general look about the code tbh
The main issue here is the setInterval on timer.js. The issue being that it was a complete bushleague move to use setInterval, instead using requestAnimationFrame would improve the way the app handles not being paid attention to it. Also rewrote some bits but its all mostly just trying to refactor code when I was riding HARD for React. 
Pull req soon tm